requirements.txt
ansible-core ~= 2.16
ansible-lint ~= 24.2
black        ~= 23.1
flake8       ~= 6.0
isort        ~= 5.12
pylint       ~= 2.17
sort-requirements ~= 1.3
yamllint     ~= 1.28
